pyproject.toml
[project]
name = "simple-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "~=3.10.17"
dependencies = [
    "aiofiles==23.2.1",
    "fastapi==0.104.1",
    "groq>=0.11.0",
    "pillow>=10.0.0",
    "pydantic==2.5.0",
    "python-dotenv==1.1.1",
    "python-multipart==0.0.6",
    "requests==2.31.0",
    "streamlit>=1.28.0",
    "uvicorn[standard]==0.24.0",
]
